Update RoomGradlePlugin KSP configuration to support KSP2
The plugin does per-variant / per-platform configuration by adding a CommandLineArgumentProvider to the KSP task, with KSP2 the task class is different and does not implement the KSP1 base class. This CL updates the plugin integration to find and also configure the KSP2 task.

KSP2 is also more strict with regards to AP options and disallows whitespaces in the key-value pairs. However it might be that a schema path might contain spaces, so the plugin and processor will now encode / decode spaces as URLs would to satisfy the format requirements.

Bug: 379159770
Test: Manually in sample project with KSP2 and Room
Change-Id: Iec3c435cbdfb994034500668f1ec602d5ebde3ce