Permit all nodes to have children detached from their lookahead
Previously, only a SubcomposeLayout's children were permitted to be
detached from their parent's lookahead pass. This was particularly
useful when the parent conditionally places the children based on
certain measurements, such as viewport and siblings' sizes.

This CL expands that support for all nodes, regardless of whether
they are in a SubcomposeLayout. This is to accomodate conditional
placements in a regular layout modifier/measure policy. Without
such support, we would *require* any placeable that may be placed in
approach to be placed in lookahead. However, such a requirement is
impossible to reinforce with the current placement APIs.
Therefore, users could easily get into an error situation that is
difficult to debug.

The downside of expanding this support to all nodes
is that the nodes involved in the make-up lookahead will not have
an accurate calculation of the future/target. But since these are
the nodes that will not exist in the future (as they are not in the
parent's lookahead pass), such the inaccurate lookahead would be
transient, therefore worth the tradeoff.

Bug: 382043554
Test: New test added
Change-Id: I279b4424b20ddce690f0cb9c870d129a11cc4b68