Implements fling gesture handling
This restores ZoomScrollGestureHandler as an inner class of PdfView. It's tightly coupled to PdfView and it ought to be an inner class.

This improves the cancellation of page link and page text fetching, and it avoids fetching links and / or text during a fling. These both mimic behaviors of the current implementation and are required to achieve reasonable latency when quickly scrolling through a document.

The tests for this required implementing Espresso's Swiper API, because of a bug in the default implementation. It's unlikely we can upstream a "fix" to the default implementation as countless tests may rely on the current behavior.

Bug: 376136621
Test: ./gradlew :pdf:pdf-viewer:(test|cAT)
Change-Id: I19a733a0334bed641cdac1da5bc9540b52c3bb9b