Eliminate dataclass DSL from ButtonGroup
Following API feedback, removing dataclass DSL from ButtonGroup.
This follows a similar pattern to core compose's ButtonGroup.
To configure items, a new modifiers are introduced in the
ButtonGroupScope: weight, minWidth, interactionSource.
Standard Material3 Buttons configure the interactionSource
automatically so it's not necesary to specify it if the Button
is a direct child of the ButtonGroup.
Relnote: Eliminate dataclass DSL from ButtonGroup
Test: Manual, passing existing tests, added one more
Bug: 370568248

Change-Id: I16c3c2c9bccc7871b337f39d36c90875396fd530