[ToggleButton] Update for design
Change the default shape to be fully rounded and the selected to be more square.
Updated toggle button samples with icon to use an icon that had more contrast between filled and outlined.
Update the outline button to remove the border when selected. Change make the outline button content transparent when unselected.

Bug: 376713020
Test: Update screenshots tests to reflect these changes.
Change-Id: I7c1d5cf9e226ec6bccb2c0f0845462c3c4f47a4e