Wait for capture session configuration if we've attempted one
In b/277675483, it was identified that on certain devices, if
CameraCaptureSession.close() isn't called explicitly,
CameraDevice.close() may block (stall) indefinitely. Consequently we
close the capture session explicitly before closing the camera device
based on previously accumulated heuristics.

There is however, an edge case that wasn't added as part of b/277675483:
When the following sequence of events occurs, we won't close the
capture session:

1. CaptureSessionState gets an opened CameraDevice
2. CaptureSessionState tries to create a capture session
3. Session has yet to be configured (we haven't received onConfigured)
4. CaptureSessionState is disconnected

In the aforementioned scenario, it is shown through lab results that, by
not closing this capture session that we've instructed the framework to
create, CameraDevice.close() actually can stall.

Here this CL adds a procedure to wait for the session to be configured
(or failed to be configured, or closed), so that we get a capture
session to close it in later sections.

Bug: 383354736
Test: VideoCaptureLifecycleStatusChangeStressTest 10/10 times on
    affected devices, and CameraX lab tests.
Change-Id: I53c5fb3be95a7328045e85509e46029633826d06