Update Confirmation to ConfirmationDialog and other API changes
Bug: 383104486
Test: N/A
Relnote: "We have updated the Confirmation API to reflect its usage as a dialog. The composable is now called ConfirmationDialog, with associated updates to the colors and defaults classes. We have also renamed the `show` parameter to `visible` for compatibility with other Compose animation APIs. In addition, we have made ConfirmationDialogContent, SuccessConfirmationDialogContent, FailureConfirmationDialogContent available for situations where developers need to customize the intro/outro dialog animations."
Change-Id: Iaeb33d3e9bb5c802ddb4b4c10ff564681772bb83