Redirect the instrinsic query to lookahead pass from a top-level lookahead root
If there is an intrinsic size query coming from above the lookahead root, we will
direct the query down to the lookahead pass. Note, when a regular measure call
reaches a top-level lookahead root, the measure call is turned into lookahead
measure followed by approach measure. This is a similar, although not exactly the
same, mental model.

This change also separates the skipToLookahead modifier into a separate file
in a mini refactor.

Test: New test added
Fixes: 382453608
Change-Id: Ic71eccd2d4eca6991bdcf451b7720ebeb24f6365