Fixes bring into view logic for reverse scrolling containers when they shrink
Previously we were trying to bring the focused child into view as part of the onRemeasured callback - but this happens before the child reports their new position in onGloballyPositioned. For reverse scrolling containers, the child position normally changes after a viewport shrink (since relative to the container it can be moved to a negative position), so if we try to bring into view with the old position, it won't work. For non-reverse scrolling containers the old position accidentally ends up working, since an item positioned at 90 in a viewport sized 100 will still be positioned at 90 when the viewport becomes 80 (for the reverse case this item could go from being positioned at 0 to -20 for example).

The existing tests in ScrollableFocusableInteractionTest did not catch this as most tests did not reverse the content layout in the reverse scrolling case, so they weren't actually testing the scenario correctly. This CL fixes the implementation, and fixes the tests to properly test the reverse scrolling case.

Fixes: b/285846153
Test: ScrollableFocusableInteractionTest
Change-Id: I9e88f144e2bc894a3d8fe02d69000cc5425524ef