Refactor JavaCompileInputs for multiplatform projects
This creates two subclasses of JavaCompileInputs, one for multiplatform projects and one for other projects.

This also renames JavaCompileInputs to CompilationInputs

Currently the multiplatform inputs is just based off of the default source set for a compilation, and matches the previous JavaCompileInputs values. The followup will add all the source sets for the compilation.

This is in preparation for creating the project structure XML file for metalava which needs information about each source set.

Bug: 326546705
Test: `./gradlew checkApi`. Tested on kzips for `:activity:activity` and `:annotation:annotation` and found no diffs (tested by running `generateKotlinKzip` before and after the change, extracting a json from the kzip with `kzip/view`, and comparing the jsons).
Change-Id: I2a1d44d0e886a37170172790e54120f024945ea3