Migrate all floatingtoolbars to use colors param
Have all the floating toolbar function using the FloatingToolbarColors
object instead of a single color param.

Update all the toolbars to use the standard color scheme by default. A
developer can choose between the standard and the vibrant, or create
their own.

Fix the padding to match the spec (8.dp) and revert a recent change that
introduced a specific ContentPadding for the toolbar with FAB. All
variations now are using the same padding.

Test: Updated
Bug: 376715181
Relnote: "Update the floating toolbar APIs to use the
FloatingToolbarColors object instead of a single container color."

Change-Id: I9a054e407d8f3518c289a1a09924ce871ad0b400