Allow comma separated list of project prefix filter
This CL allows PROJECT_PREFIX filter to be a list of prefixes
instead of one prefix.

This is particularly useful when working on coupled projects
(like room and sqlite). For that scenario, just using room
is not sufficient because even though it depends on sqlite, it
does not depend on sqlite's integration tests.

Bug: n/a
Test: PROJECT_PREFIX=:room,:sqlite ./gradlew projects

Change-Id: Ia20156fa36fd18fb8e22812fe62e3140e2a5ba36