Add SavedStateHandles for ViewModelNavContentWrapper
Adding the ability to use SavedStateHandlers with ViewModels created
from the ViewModelNavContentWrapper that are record scoped.

Normally the ViewModelStoreOwner and SavedStateRegistryOwner creating
the SavedStateHandle are the same component. Since in this case they are
different, we need a way for the `enableSavedStateHandles` call to be
timed before the registry's controller is attached, but the controller
to be attached before the content containing the ViewModel is composed.

To do this, we can move our performRestore and enabledSavedStateHandle
calls to the inits of the SavedStateRegistryOwner and
ViewModelStoreOwner respectively. This way SavedStateHandles are enabled
before we ever reach composition.

RelNote: N/A
Test: Added ViewModelStoreNavContentWrapper
Bug: 380445098
Change-Id: I0d47778e02dd16286887c073e6011e838ddec03a