Remove ambiguous links to `alignBy` in `RowScope` documentation.
Kdoc bracket-link syntax doesn't include parameters and is ambiguous when multiple resolution candidates of the same type (function, classlike, etc.) occur in the same context.  
In particular, this can cause nondeterministic output, which interferes with our hermetic build enforcement and pure-refactor no-op detection.

These change should not be a problem because most of the changed links were self-links, and the others were from `alignByBaseline` which occurs immediately next to `alignBy` on the docs pages.

Change-Id: Ic0a9b2685ed5bf3c8418a65a89cb2b0863d81904
Test: documentation-only change.