Avoid thread switches when getting SemanticsNodes
TestOwner.getAllSemanticsNodes switches to the main thread in order to
call SemanticsOwner.getAllSemanticsNodes. When it returns the result (on
the test thread), every caller continues with switching back to the main
thread to process the result. This is inefficient and could lead to race
conditions. Solve this by adding a transformation callback to
TestOwner.getAllSemanticsNodes so that processing can be done in the
same main thread execution.

Fixes: 376892469
Test: presubmit
Change-Id: Ie17774775b77a65e8195ef8bb50714fd6f2cf407