Change the condition to select nodes related to autoCommit
We now use the presence of a contentTag as a signal that the node is related to autocommit.
We call notifyViewEntered only if the focused item is autofillable.
We populateViewStructure for nodes that have any semantics that are related to autofill.

Also updated some tests which set focus semantics manually instead of
actually changing focus. Removed some helper modifiers in tests by
using Modifier.then instead.

Test: Updated existing tests.
Change-Id: I750e2e462557afb0824cc0bde15ab45aecc2fec2