Clean up usages of "field" in KDocs & generated code.
This CL refactors all references to "field"s to "property" since we should be using KSP-compatible terminology in
the KSP version of Room.

This CL also cleans up any leftover mentions of POJOs and "method" in the KDocs that were not included in the initial refactor CLs.

Bug: 364902863
Test: Existing tests.
Relnote: Refactor all references to field to property in KDocs & generated code in Room.
Change-Id: I1e111a20717db93481f592a89798731df87d5679