Remove auto centering logic from SLCScrollInfoProvider
Details in bug report. Auto centering logic here was causing anchor
item to be set to an offscreen item in SLC, which meant scroll away
was not triggered for time text at start of list, instead it got
triggered further down the list when the auto centering item index
became visible, which caused the time text being stuck issue.

After this change the anchorItemOffset is calculated by looking at
a fixed position item near the top of the list (item 1), since this
anchorItemOffset is primarily used to determine when to scroll away
time text and item 1 is likely always visible when the SLC is first
displayed, this is an appropriate index to use. 0 can also probably
be used, however since 1 is already the default value when auto
centering is not used it is probably better to stick to this.

Bug: 372750873
Test: Manual (video included on bug)

Change-Id: Ie6a49cdaa995fe71bab69f4f4d62e79923fd78a5