Seal PaneMotion interface and remove its default transition methods
To address API council feedback. Also it gives us the ability to support
offset animation when entering with expanding and exiting with
shrinking - originally PaneMotion has no idea about its own role and
where the associaetd pane is; after the change, the transition will be
created from the motion data provider with the given role, which
provides more info when creating the transition, and in the meantime
also ensures the behavioral consistency better as we can ensure the
involved PaneMotion is associated to one of the panes.

The change also encourages clients directly customizing enter/exit
transitions from the respective paramters introduced to the AnimatedPane
function.

Relnote: "Seal PaneMotion interface and remove its default transition
  methods."

Test: existing tests
Bug: 381921988
Change-Id: Ifc4c783ba071091ee5025a57943f6ae4dc82d72d