Introduce Offset.fromStart() and Offset.fromEnd() factory functions
According to the UXR feedback we receive, the original API design that
uses negative value to represent offsets from the end is unclear and can
be confusing. Also it's impossible to represent a zero offset from the
end in the original design.

Changes the API design to:

1. Hide the constructor.
2. Introduce two factory functions fromStart() and fromEnd() so it's
   clear how to set an offset from the end.

Relnote: "Introduce factory functions to create pane expansion anchors
  of the offset type from the start and the end, respectively."
Test: existing unit tests
Change-Id: I2eb21917bbc9b63fb047eaf42f57784b4a54da06