Fix fillInLabel not encoded by NavType
When parsing labels in XML, the label will be retrieved via NavType.get. This means we will now respect the custom logic on how the label is to be retrieved from the SavedState.

Test: ./gradlew navigation:navigation-ui:cC
Bug: 388100796
Relnote: "NavDestination labels provided through XML will be parsed via NavType.get to ensure custom NavType logic is respected."
Change-Id: I7ec2f3e410b616e8933c8f5dd8015869906fb263