Change the nullability of setOnApplyWindowInsetsListener
Since setOnApplyWindowInsetsListener overrides a method from the View
class which is in java and has no nullability, we did not know the
proper nullability.

It seems there are cases when the inset listener could indeed be null,
so we can make that nullable now.

RelNote: "FragmentContainerView's `setOnApplyWindowInsetsListener`
override now takes a null listener."
Test: All tests pass
Bug: 282790626

Change-Id: I575f0b5da72d096b14393afd9819ab66e061b002