Include minLines in the calculation of the intrinsics height
maxLines is included in the intrinsics height calculation because on measure we pass the maxLines to the platform. minLines can't be pass as an argument to the platform's measure object.
So instead we coerce the platform's measured intrinsics height so it's at least as small as the height required to satisfy the minLines.

Fixes: 388299762
Relnote: "Text's minimum intrinsics height now takes `minLines` parameter into account. It means that minimum intrinsics height reported won't be smaller than the height required to satisfy the `minLines` parameter"
Test: new test minIntrinsicsHeight_respectsMinLines + demo
Change-Id: I225f910dcbff2292d045ffac0750544d4601eeba