Add a more customizable TwoRowsTopAppBar
Add API for a more generic two-rows top app bar to allow more
flexibility for creators.

The API has slots for the title and subtitle in their expanded and
collapsed states, navigation icon, and actions.
The sizing and styling of the text in those slots is defaulting to what
the MediumFlexibleTopAppBar is applying, so creators may override that
by wrapping their content in a composition local. The expanded and
collapsed heights are also defaulting to the medium top app bar heights,
so a creators may provide custom size through their dedicated
parameters (`collapsedHeight` and `expandedHeight`).

Bug: 306697446
Bug: 229134133
Bug: 268068946
Test: added
Relnote: "Added TwoRowsTopAppBar API, a more customizable alternative
to Medium and Large app bars for creating two-row top app bars.
Developers now have control over the content of titles and subtitles
in both collapsed and expanded states, with customizable heights for
each state."

Change-Id: I0be3cd4f82ca3d0e74ea8b5dc0f2dcc0cea43e52