Add an empty consumer proguard file if none was specified
In order to prepare for R8 automatically shrinking libraries we want to
mark all libraries that don't yet have a proguard consumer file with
a blank one, so R8 know it is safe to shrink this library.

This is a follow up for r.android.com/3445723 and applies to AARs
in KMP projects.

Bug: 388837472
Test: Unable to test due to b/388938580
Change-Id: If7fbd233688307202eac18307daf866b267dbf7c