Use RequestCloseById for UseCaseManager and Camera2CameraController
Prior to b/324142928 and b/388054983, there was no way for users to wait
for the camera to be closed, nor does the close request take priority.
This can lead to issues with correctness, and thus we simply opt to
close all cameras during shutdown. Now that RequestCloseById is done
with proper pruning strategies, and we have the ability to wait for it,
we can opt to use it rather than closing all cameras.

Bug: 341410063
Bug: 324142928
Bug: 388054983
Test: CameraX test lab
Change-Id: Ia177b1bf5a70dde6cd3d33bdf207ead320d71ef1