Fix: Crash caused by IllegalArgumentException when unbinding a service that's not registered.
The crash happens because of a race condition where the disconnect function is called asynchronously. There’s a delay in updating the mConnected variable after the service is disconnected, so another call to disconnect tries to unbind the already disconnected service, causing the error.

The disconnect function now uses locks to ensure it runs synchronously, preventing race conditions and handling the service unbinding safely.

Bug: 388160048
Test: build and run on test device
Change-Id: I7448cdc70ccefb96fb47ba6c0e736a2bf217e1b2