Remove generic from CharSequenceSerializer
`Parcel` on Android does not retain the actual type of `CharSequence` so in this CL we removed the generic in `CharSequenceSerializer` and made it final.

I also changed the actual type from `String` to `StringBuilder` in our tests on `CharSequence` to make it obvious that the `CharSequence` restored from `Parcel` is no longer a `StringBuilder`.

Relnote: Remove generic from CharSequenceSerializer
Test: SavedStateCodecAndroidTest.android.kt
Change-Id: Ib40bd2dcf522ceb14f03efacffc18ce6ad188844