Add platform-specific encoding and decoding to SavedState codec tests
Add platform-specific encoding and decoding (parcelling and unparcelling on Android and no-op on other platforms) to SavedState codec tests for increased coverage.

After this change, we got `ClassCastException` when restoring arrays of sub-type of `Parcelable`s. This is because the array restored is of type `Array<Parcelable>` and we cannot assign it to an array of Parcelable's subtype, e.g. `Array<MyParcelable>`. I haven't found a way to create instances of `Array<subtype of Parcelable>` in our serializer so I changed the tests to reflect that using `ParcelableArraySerializer` with `Array<subtype of Parcelable>` will lead to `ClassCastException`. Now in all our built-in serializers for collection types, only `ParcelableListSerializer` and `SparseParcelableArraySerializer` can be used with subtypes.

A `checkDecoded` parameter is added to `encodeDecode()`. This can be used to check that the decoded Serializable is essentially the same as the original Serializable when its `equals()` doesn't work.

Bug: 387270265
Test: SavedStateCodecAndroidTest.android.kt↲
Change-Id: I31beee5105f1a6a5495e24f6c9e643e8550bc6cc