Ignore ImageWriter creation failures
When a camera graph with a reprocessing configuration is shut down
immediately after creation, there is a chance we have a configured
session after the user already releases the input Surface. In such
cases, we'd be creating an ImageWriter to that released Surface, causing
an exception. Catch the exception here for the aforementioned case, and
if it isn't, and the user still tries to submit a reprocessing request,
we would then check that and throw.

Bug: 378129259
Test: CameraPipe test app
Change-Id: I905259581e6b632dc26e9664730ccfb2e01df98c