[GH] room-paging refresh initial load fix
## Proposed Changes

If items are removed then the initial key for a refresh may be close enough to the end of the list that fewer than a screens-worth of items may be loaded causing the UI to jump. Changed to handle similarly to the case where the initial key larger than itemCount and load loadSize items from the end.

## Testing

Test: ./gradlew test connectedCheck

## Issues Fixed

Fixes: b/389729367

This is an imported pull request from https://github.com/androidx/androidx/pull/750.

Resolves #750
Github-Pr-Head-Sha: e01574c59eaeb92cffd0cd1606b213b7fc5af475
GitOrigin-RevId: a6c2d92a4012757817fd74af4fc1a57fb6de8f88
Change-Id: I2abbe92633bf95aa3306e4c2ebc47eacebfffa8c