Fix focus move within child AndroidViews
Fixes: 389994198
Fixes: 391378895

Relnote: "Fixes focus travel within child AndroidViews where it
can sometimes jump out of the child to the ComposeView"

Sometimes the native FocusFinder doesn't act properly and finds
the wrong View when searching. This CL changes the FocusFinder to
use a newer copy of FocusFinder for all focus search operations
in AndroidComposeView.

Test: new test, manual testing
Change-Id: I678c864afe086879c5c45b018dd82da4f23be440