Add `retain { ... }` API
This CL adds `retain { ... }` to the Compose Runtime. The retain API
is similar to remember in that it persists values across recompositions,
but retained values are also able to persist destruction and recreation
of the composition hierarchy. This gives them similar scoping to a
ViewModel or Fragment.

Currently, the default installed scope is the ForgetfulRetainScope,
which makes retain behave the same as remember. Follow-up CLs will
install an Activity-lifecycle bound scope and other use-case specific
RetainScopes (like navigation) in our 1p androidx libraries.

Design doc: go/remember-retained-proposal
Bug: b/177562901
Test: RetainedTests
Relnote: """
    Introduces the retain API. Like rememberSaveable, retain can
    persist and save remembered values that leave the composition
    hierarchy. Retained values are not serialized and have a shorter
    lifespan than saved values, and behave similarly to an androidx
    ViewModel. See the documentation for more details.
    """

Change-Id: Ia31056d0ea23d6358a11d8ad9f472b51f41879d8