Make JavaSerializableSerializer and ParcelableSerializer abstract
This forces users to extend these two serializers in order for them to be used with JavaSerializable, Parcelable, or their subclasses. For example, this used to work but not anymore because Kotlin Serialization can not create the abstract ParcelableSerializer:

```
@Suppress("SERIALIZER_TYPE_INCOMPATIBLE")
@Serializable(with = ParcelableSerializer::class) val value: MyParcelable
```

Instead, we ask users to extend these serializers like this:

```
class MyParcelableSerializer : ParcelableSerializer<MyParcelable>()
```

And then it can be used normally. Note that no SERIALIZER_TYPE_INCOMPATIBLE suppression is needed:

```
@Serializable(with = MyParcelableSerializer::class) val value: MyParcelable
```

Test: SavedStateCodecAndroidTest
Relnote: Make JavaSerializableSerializer and ParcelableSerializer abstract
Change-Id: I268f62533803b846f158b51a65da091285f90dd3