Increase compositeKeyHash precision
As explained in go/remember-retained-proposal, our current
compositeHashKey has a relatively high probability of two groups
colliding due to the birthday problem. It takes as little as ~77,000
groups in the SlotTable before any two groups chosen at random have a
50% chance of having the same hash key. This statistic assumes a perfect
hash, which the composer does not have, so in practice it takes many
fewer groups before there is a 50% chance of hash collision.

By expanding the width of our key from a 32-bit int to a 64-bit long
(Or a 52-bit number on JS), we exponentially reduce the chances of a
collision. On the JVM, the maximum number of groups that the current
SlotTable is theoretically capable of addressing is around 429 million
groups. If such a SlotTable were constructed, it would consume 8GB of
memory just for the group metadata, and have less than 0.5% chance of
two groups having the same hash if it is stored with a 64-bit long
(assuming a perfect hashing function).

In practice, collisions are still possible, because the hash is not
perfect and there are a few known cases where the hash will produce the
same values.

Test: CompoundHashKeyTests
Bug: b/177562901
Relnote: """
    currentCompositeKeyHash is now deprecated. Use
    currentCompositeKeyHashCode instead. The replacement API encodes the
    same hash with more bits, which exponentially reduces the chance of
    two random unrelated groups in the composition hierarchy from having
    the same hash key.
    """
Change-Id: I4cb6a801d12447ac52bc92bb899ae84d69c4a6ed