Allow permission checking fallback
In b/366201463 we modified obtainAndCheckReceiverPermission to obtain a
different permission for API levels 29+. On some OEM builds, this
appears to be causing problems, while fixing problems on others. To
handle both cases (it is indeed possible that the OEMs are doing the
incorrect thing), we will check both permission as a fallback, and in
the thrown error, indicate the correct permission. Since this is adding
an additional case in which not to throw an error, this is low risk, and
will reduce the number of crashes.

Bug: 366210463
Test: not possible
Change-Id: I18eb20b2683adada6973724877f839155a727359