Implement popTransitions in NavDisplay
This implementation only applies to non-Dialog NavEntries. The implementation for Dialog will follow.

The backstack as a List means users can manipulate the backstack at any position. They can make changes in the middle without changing the top-most entry. Or they can change entries at random positions with an actual pop at the end.

As a ground rule in this implementation, a backstack operation is treated as a pop only if the newBackStack is a proper subset (subset but not equal) of oldBackStack.

Test: ./gradlew navigation3:navigation3:cC
Bug: 389974333
Relnote: "Users can now apply animations when popping NavEntries by providing popTransition through NavDisplay or NavEntry.featureMap."
Change-Id: I22077a92037e759e3364a59d0143a51951f89fca