Support touch passthrough feature in Custom Tabs
The new API ActivityOptions#setAllowPassThroughOnTouchOutside is
being applied to CustomTabsIntent#Builder so that apps opening
partial-height Custom Tabs can control the touch event passthrough
settings. This was already possible on pre-B via pCCT manually
filtering the events since it is enabled by default, but B disables
it for security reasons. On B+ this feature will be controlled
via the new API.

For apps, the usage remains the same as the config is done with the
same intent extra flag EXTRA_DISABLE_BACKGROUND_INTERACTION.
CustomTabsIntent#Builder switches to using the new API internally
on Android B+.

Test: manual - ran on ZP1A using a test app with target SDK 10000
Bug: 370100158

Change-Id: Ie84f331be730bdae083ac5f7a3f0af72a9adbf89