Clarify test for rememberSaveable
Moved the check for the changed integered to before the new key is added
since there is no guarantee whether we will get a composition or during
the animation. There is no guarantee that compose will skip the
compostion of the previous content associated with the previous key so
this change avoid relying on that and just ensures the state is restored
properly when the content returns.

Test: changes test
Change-Id: I2a2becc7111771557cfcfe89ab011351529b973e