Fix SupportedSurfaceCombination NPE issue when extensions supports additional JPEG_R format
On some devices, when extensions is on, some extra formats might be supported for extensions. But those formats are not supported in the normal mode. In that case, MaximumSize in SurfaceSizeDefinition could be null. Directly make the ConfigSize to be MAXIMUM when the input size is larger then RECORD size.

Bug: 388141689
Test: ImageCaptureTest#canBindToLifeCycleAndTakeJpegUltraHdrPicture on pixel-9-pro
Change-Id: Ide4c31e891a5a412e97781ff9f3e9f6f33b48eba