Ensure ActiveCamera canoot be connected with WakeLock closed
The current ActiveCamera mechanism relies on a WakeLock to enact a 1s
timeout to close the camera after a VirtualCamera is disconnected. When
the timeout expires, the WakeLock is synchronously closed, and the
(close) callback is invoked, which sends a RequestClose request to the
device manager.

The problem here is, until that RequestClose request is processed, the
ActiveCamera stays in the available active cameras in the device
manager, and it can still be opened. This makes it possible for a
RequestOpen to get an ActiveCamera with its WakeLock already closed.

Another minor problem is, while we're processing a RequestOpen request,
we may fetch an ActiveCamera with a not-yet-closed WakeLock, but then it
can actually still be closed before we connect it to the VirtualCamera,
since the WakeLock operates under its own coroutine scope.

The key to fixing this issue is realizing that the success of an open
request, really hinges on the status of the WakeLock. An ActiveCamera
can only be used if we can acquire a token from its WakeLock. With this
understanding, this change does the following:

- Separate ActiveCamera.connectTo(VirtualCameraState) into
  ActiveCamera.acquire() and
  ActiveCamera.connectTo(VirtualCameraState, Token)
- When we're fetching an ActiveCamera for an open request, also acquire
  its Token, essentially marking it as used. This also then prevents
  WakeLock from closing itself in parallel.
    - If we're unable to acquire the Token, that means the WakeLock
      closure won the race, and we should reopen the camera.

This CL also includes a functional behavior adjustment for concurrent
cameras. Previously if you were to get a pair of CameraGraphs for
concurrent cameras and you didn't start both within 1s, it may not
succeed. This is in hindsight not a great behavior, since by starting 1
CameraGraph, it effectively means the camera should be held. Generating
2 CameraGraph configurations may also take more than 1s as another
thought.

Bug: 390530866
Test: CameraX test lab and VideoRecordingTest 20/20 times on affected
  devices.
Change-Id: I4972b0e930eed2c92ffa5546e625e3e2bb31e69a