Don't close the session if SandboxedSdkUi is in LazyList
SandboxedSdkUi must keep the session open when it is detached and re-attached while inside Lazy list.
Session management logic in SandboxedSdkView does not work out of the box for SandboxedSdkUi
as Lazy List does not implement PoolingContainer API of view framework.
Session is kept open if node can be re-used and is closed when it is relased and will not be used anymore.

Bug: 340466873
Test: Manual
Change-Id: I4d26cb7b150d7dc6f94905e84ce04dc4fdc9bc21