Add SLC benchmark to M3 benchmarks and refactor TLC benchmark.
Add ScalingLazyColumnBenchmark macrobenchmark to Material3
macrobenchmark, and refactor both TLC and SLC benchmarks to use
device.swipe() instead of list.drag(). The change makes it easier to
compare SLC, TLC and Picker Benchmarks.
Also increased number of swipes in the list/picker benchmarks from 4 to
20, for bigger sample size and more reliable results.

The results for benchmark runs are attached in the bug.

Bug: 389878201
Test: Update benchmark tests
Change-Id: I6bad9e993e6ef9316fc6d129af0388d28cb386a6