Fix MinLinesConstrainer caching issue
We cache a previously computed MinLinesConstrainer. When we check if the params changed so we can return the cached value, we compare style with which the "from" is called with the styles that the cached constrainer is created with. The only issue is that when we create a constrainer, we resolve the defaults in ctor. So we end up in a situation where we always compare raw style with resolved defaults style resulting into cache miss even when the style that the `MinLinesConstrainer.from` is called is identical.

Bug: 388299762
Test: new MinLinesConstrainerTest
Change-Id: Ie5237548a60b154b6e927406cc108ee7bc7432c6