Fix out of order execution with StartRepeating during AbortCaptures
In GraphLoop, when AbortCaptures is processed, the captures are aborted
first, then a sweep is conducted to remove preceding, nullified
elements. However, StartRepeating is not removed as part of the process.
This means if there is a sequence, like [`StartRepeating`,
`AbortCaptures`]. `AbortCaptures` will be executed, while leaving
`StartRepeating` to be executed later, while `AbortCaptures` should've
cancelled the `StartRepeating` element.

Bug: 391467026
Bug: 359338491
Test: Testing
Change-Id: If3d7e5f31ba0ce744bd92b1d35fdd351c1ece34d