Drop shared elements from overlay when transition ends
Previously shared elements transition assumes only one shared element
among all the matches will be in the tree once the shared
element transition is finished. But that's not the case when
using `sharedElementWithCallerManagedVisibility`, where both
instances of shared elements will co-exist even after the transition.
Hence the logic for whether shared elements should render in the
overlay needs to also account for whether the transition is
ongoing, in order to drop shared elements from the overlay
once the transition ends.

Fixes: 389951209
Test: New test added
Change-Id: Ic9f18c4ee93be9e3fad2c80950ac0d2eee8961da