Replace NavWrapperManager with NavBackStackProvider
Adding a new NavBackStackProvider that handles required operations
automatically. More specifically, it makes the proper calls on all of
the NavLocalProviders and then provides a list of NavEntries whose
content is completely ready to be invoked.

This changed the ProvideToBackStack API to take and provide a
contentLambda and made the ViewModelStoreNavLocalProvider remember its
owner to ensure it does not get recomposed unnecessarily.

RelNote: n/a
Test: Modified tests
Bug: 391683835
Change-Id: I8024395feb7708a5458498315dcbb02eb9282501