OpenOnPhoneDialog: Address api council feedback regarding `show`, `curvedText` and other parameters.
Renamed the `show` parameter to `visible` in the `OpenOnPhoneDialog` composable for better clarity and consistency with other Dialogs. Also, the curvedText content is now provided by the caller instead of having a default value. This allows more customization and flexibility in how the dialog is displayed.

Bug: 385132997
Bug: 389659709
Test: Updated existing tests.
Relnote: "OpenOnPhoneDialog api was updated for better clarity and consistency with other Dialogs. `show` parameter was renamed to `visible` and `curvedText` is now provided by the caller instead of having a default value."
Change-Id: Idec2d959e1a9c3bab7ad193c159d68be8a23c194