Improve sync trigger transaction avoidance.
The commit Iae612a1c4f8b8f1e8e15d2f60eb9bf68241570ef attempts to avoid a transaction if no trigger needs to be added or removed but it did so by using `any()`` which O(N). If we move the logic of detecting that no add or remove is needed to be in the same iteration that is building the operation list we can entirely avoid the `any()` check.

Bug: 391962461
Test: ObservedTableStatesTest
Change-Id: Icc84fb31df94454bbb79b80a8dcd9e836e66581f