Add ui-test-accessibility and ui-test-junit4-accessibility library
This allows us to remove the hard dependency on ATF inside ui-test and ui-test-junit.
Methods like enableAccessibilityChecks() and disableAccessibilityChecks() on ComposeUITest and ComposeTestRule moved to those packages correspondingly.
It also addresses the API usage feedback and adds accessibilityValidator as a parameter to the enableAccessibilityChecks which allows to easily configure and pass your own ATF validator.

Given that Android is the only platform right now supporting these checks, we made this API Android-only. Inside ui-test everything is ready to work on various platforms. If we decide to move the enableChecks APIs in these packages to KMP, we'll likely deprecate hidden the existing and add the new one, perhaps tweaking the jvm name if needed.

The integration point between the ui-test and ui-test-accessibility is a ComposeAccessibilityValidator that is made RestrictTo to avoid exposing API added in such short notice. In the future we can consider making it public.

Most of the tests, samples and documentation are moved as is adjusting for a slightly changed API.

Test: AccessibilityChecksTest
Test: verified running checks in a separate (foundation) module
Fixes: 391948233
Fixes: 391560768
Relnote: "Moved `enableAccessibilityChecks()` API that turns on the Accessibility Checks for Android to separate packages which are `compose:ui:ui-test-accessibility` if you're not using the TestRule and `compose:ui:ui-test-junit4-accessibility` when calling on TestRule"

Change-Id: I3c3183c7f06ced8c0ad4873b6262ea39e1ea1b16