Add errorDim to the Material3 ColorScheme
UX requirement to add additional error token, for both apps and tiles.
UX is working on defining use cases to be put in Javadocs for
developers.

Based on Compose impl, aosp/3460282

Since this token doesn't map to platform tokens (that were already
finalized), we need to do calculations.

Fixes: 392803988
Test: Locally. Will update samples APK for UX review
Relnote: "We have added errorDim to the Material3 ProtoLayout ColorScheme, for high priority errors or emergency actions such as safety alerts."
Change-Id: Ia17bbb91fbe327cc4ac0e57ba22c5cf2b445fac4