Add support for an edge-swipe zone to SwipeToReveal.
Bug: 381403787
Test: Unit tests added.
Relnote: "Add support for an edge-swipe zone to SwipeToReveal. Foundation
SwipeToReveal changed to, by default, not allow it to be swiped when the
gesture starts from the edge. Material3 SwipeToReveal changed to, by
default for when SwipeDirection is set to single direction, not allow it
to be swiped when the gesture starts from the edge."

Change-Id: I32ef0e9e201fc71cb5c9830c766a933a84b99c74