Support Compose stack traces in LayoutNode phases
This change adds a composition local that exposes access to the slot table lookup through `CompositionErrorContext`. The layout node now rethrows exceptions in measure / layout / draw with a Compose stack trace attached.

Test: UiErrorTraceTests
Bug: 354163858
Relnote: "Expose a composition local that allows to attach a compose stack trace based on a compose node location in composition."
Change-Id: Ie0bdaffb43a46328057c31e7fc9db51a80789b5e