Handle platform type converters in Kotlin codegen
In Room platform types are treated as nullable so they an be properly bound and read from statements. In KSP with Kotlin codegen, platform types from converters should go through a null check converter but instead it was though a up-cast. This CL changes the behaviour on the converter store to determine assignability treating platform types as nullable.

Bug: 339809512
Test: CustomConverterProcessorTest
Change-Id: I6d3909815e953ef3e70b38b15059d1d6d9e4881f