Add FontSetting parameter to Material3 text method
After UX raised, not all numerals use-cases should use tnum feature.
Consequently, we should allow developers to be able to override default
values for these settings.

How setting behave when axis is not supported will be treated in a
follow up, as in general Android mostly ignores the axis.

Additionally, fix default value for scalable, because it shouldn't be
read from Scope as default depends on which typography is used, so text
will set it directly.

Fixes: 393176383
Test: Added
Relnote: "Add FontSetting list parameter for Material3 text."
Change-Id: Ic102d8e67624b4379690dc6144f5f0a164ae60f4