Allow handling of non-consecutive duplicate keys
This change simplifies how the SaveableStateNavLocalProvider handles
saving and restoring the state from the SaveableStateProvider.

Now ProvideToBackStack is solely responsible for tracking the state
associated with the key since whether a key is in the backstack is what
determines when it should be removed.

ProvideToEntry only retrieves and provides that state to the entry's
content. It has no determination on how long that state should survive.

The count of a specific key in the back stack at any given time is also
considered which allows reusing the same key as we can apply different
state.

Test: Added test with duplications
Bug: 391607910
Change-Id: I9490707522f41aeeb54a2c32aee2af2f2d48b094