Fix ImageAnalysisTest _withRotationEnabledAndReusedToHaveDifferentSize tests failures
analysisResults is cleared too early. Some images might still be recieved after it is cleared. Update the code to clear it right before resetting the analyzer to ImageAnalysis. Also, verify the result with the last item in case the first item is still possible to be of the previous resolution because it took some time to complete the format conversion process.

Bug: 392856127
Test: ImageAnalysisTes on some lab detives
Change-Id: Ia36a5fb185f515f3b2b04ee08e2f18c91d7c6d63